backend/app/config.py
import os
from pathlib import Path

from dotenv import load_dotenv

load_dotenv()

BASE_DIR = Path(__file__).resolve().parent.parent


class Settings:
    groq_api_key: str = os.getenv("GROQ_API_KEY", "")
    database_url: str = os.getenv("DATABASE_URL", "sqlite:///./digilocker.db")
    upload_dir: str = os.getenv("UPLOAD_DIR", "uploads")
    max_file_size_mb: int = int(os.getenv("MAX_FILE_SIZE_MB", "10"))
    processing_timeout_seconds: int = int(os.getenv("PROCESSING_TIMEOUT_SECONDS", "20"))


settings = Settings()
