backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pymupdf
easyocr
groq
python-dotenv
pillow
aiofiles
